api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic-settings==2.5.2
python-dotenv==1.0.1
alembic==1.13.2
aiogram==3.13.1
httpx==0.27.2
